refactor: remove deprecated ChatRoom page
