+++ src/router/index.js
@@ -7,7 +7,6 @@ import BlessingView from "@/views/BlessingView.vue";
 import CartView from "@/views/CartView.vue";
 import ChatRoomView from "@/views/ChatRoomView.vue";
 import EventCardByIdView from "@/views/EventCardByIdView.vue";
-import ChatRoomView_new from "@/views/ChatRoomView_new.vue"
 import EditProfileView from "@/views/EditProfileView.vue";
 import GiftCheckoutView from "@/views/GiftCheckoutView.vue";
 import HomeView from "@/views/HomeView.vue";
@@ -82,14 +81,9 @@ const router = createRouter({
       component: () => import("@/views/EventCardByIdView.vue"), 
     },
     {
-      path: "/chat_new",
-      name: "ChatRoomView_new",
-      component: ChatRoomView_new,
-    },
-    {
-      path: "/chat_new",
-      name: "ChatRoomView_new",
-      component: ChatRoomView_new,
+      path: "/chat",
+      name: "ChatRoomView",
+      component: ChatRoomView,
     },
     {
       path: "/edit-profile",
